basecode/pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
dependencies = [

    "toml>=0.10.2",
    "locust>=2.14.0",
    "pexpect>=4.8.0",
    "pytest-aiohttp>=1.0.4",
    "selenium>=4.7.2",
    "Appium-Python-Client>=2.7.1",
    "webdriver-manager>=3.8.5",
    "typer>=0.7.0",
    "rich>=13.0.0",
    "pytest>=7.2.0",
    "pytest-html>=3.2.0"
]

name = "basecode"
version = "1.0.0"

[tool.setuptools.packages.find]
where = ["basecode"]

[tool.setuptools.package-data]
"*" = ["pyproject.toml", "**/pytest.ini"]

[project.scripts]
basecode = "basecode.src.__main__:main"
